refactor: split page behavior scripts
